Move platform-specific files into their own folders
