+++ test/Microsoft.Azure.Devices.Edge.Test.Common/IEdgeDaemon.cs
@@ -1,3 +1,4 @@
+// Copyright (c) Microsoft. All rights reserved.
 using System;
 using System.Threading;
 using System.Threading.Tasks;
